src/Modules/Garuda.Modules.ProjectManagement/Controllers/BankController.cs
// <copyright file="BankController.cs" company="CV Garuda Infinity Kreasindo">
// Copyright (c) CV Garuda Infinity Kreasindo. All rights reserved.
// </copyright>

using System.Collections.Generic;
using System.Threading.Tasks;
using Garuda.Infrastructure.Constants;
using Garuda.Infrastructure.Dtos;
using Garuda.Modules.ProjectManagement.Dtos.Requests.Bank;
using Garuda.Modules.ProjectManagement.Dtos.Responses.Bank;
using Garuda.Modules.ProjectManagement.Services.Contracts;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace Garuda.Modules.ProjectManagement.Controllers
{
    [Authorize]
    [Produces("application/json")]
    [Route("api/banks")]
    public class BankController : Controller
    {
        private readonly IBankServices _bankServices;

        /// <summary>
        /// Initializes a new instance of the <see cref="BankController"/> class.
        /// </summary>
        /// <param name="technologyServices"></param>
        public BankController(IBankServices bankServices)
        {
            _bankServices = bankServices;
        }

        /// <summary>
        /// Get list of banks.
        /// </summary>
        /// <returns>A <see cref="BankResponses"/> representing the asynchronous operation.</returns>
        [HttpGet]
        [ProducesResponseType(Codes.SUCCESS, Type = typeof(MessageDto))]
        [ProducesResponseType(Codes.NOT_FOUND, Type = typeof(MessageDto))]
        public async Task<IActionResult> GetListBank()
        {
            var result = await _bankServices.GetListBank();
            return Ok(result);
        }

        /// <summary>
        /// Create new bank
        /// </summary>
        /// <returns>A <see cref="MessageDto"/> representing the asynchronous operation.</returns>
        [HttpPost]
        [ProducesResponseType(Codes.SUCCESS, Type = typeof(MessageDto))]
        [ProducesResponseType(Codes.NOT_FOUND, Type = typeof(MessageDto))]
        [ProducesResponseType(Codes.BAD_REQUEST)]
        public async Task<IActionResult> CreateBank([FromBody] CreateBankRequests model)
        {
            if (!ModelState.IsValid)
            {
                return BadRequest(ModelState);
            }

            var result = await _bankServices.CreateBank(model);
            return Ok(result);
        }
    }
}
